[LV] Move VPlan-based calculateREgisterUsage to VPlanAnalysis (NFC).
